+++ runner.go
@@ -37,8 +37,7 @@ func heartbeat(mgr *Manager) {
 			// we don't care about errors, assume any network
 			// errors will heal eventually
 			err := mgr.with(func(c *faktory.Client) error {
-				//data, err := c.Beat(mgr.CurrentState())
-				data, err := c.Beat()
+				data, err := c.Beat(mgr.state)
 				if err != nil || data == "" {
 					return err
 				}